cybcash의 자료형 long으로 변경
Update src/main/java/com/kh/bts/acnt/model/vo/Acnt.java
Update src/main/java/com/kh/bts/admin/controller/AdminCtrl.java
Update src/main/java/com/kh/bts/admin/model/dao/AdminDao.java
Update src/main/java/com/kh/bts/admin/model/service/AdminService.java
Update src/main/java/com/kh/bts/admin/model/service/AdminServiceImpl.java
Update src/main/java/com/kh/bts/investment/controller/investmentCtrl.java
Update src/main/java/com/kh/bts/member/model/vo/Member.java
Update src/main/java/com/kh/bts/mypage/controller/MypageController.java
Update src/main/resources/mappers/acnt-mapper.xml
Update src/main/webapp/WEB-INF/views/admin/afterReplyReport.jsp

+++ src/main/java/com/kh/bts/acnt/model/vo/Acnt.java
@@ -6,14 +6,14 @@ import org.springframework.stereotype.Component;
 public class Acnt {
 	private String acntno;
 	private String email;
-	private int cybcash;
+	private long cybcash;
 	private int bankPw;
 
 	public Acnt() {
 	}
 
 
-	public Acnt(String acntno, String email, int cybcash, int bankPw) {
+	public Acnt(String acntno, String email, long cybcash, int bankPw) {
 		super();
 		this.acntno = acntno;
 		this.email = email;
@@ -30,11 +30,11 @@ public class Acnt {
 		this.acntno = acntno;
 	}
 
-	public int getCybcash() {
+	public long getCybcash() {
 		return cybcash;
 	}
 
-	public void setCybcash(int cybcash) {
+	public void setCybcash(long cybcash) {
 		this.cybcash = cybcash;
 	}
