+++ src/main/java/com/kh/bts/admin/model/service/AdminService.java
@@ -23,7 +23,7 @@ public interface AdminService {
 	public List<CashLog> searchCashLog(String keyword);
 	public List<Member> adminSearchMember(String keyword, int searchType);
 	public List<Member> adminListMember(int startPage, int limit);
-	public int totalCybcash();
+	public long totalCybcash();
 	public int countTodayWon();
 	public int countCreport();
 	public List<Creport> selectCreport(int startPage, int limit);
